CENTURI Hackathon:Flytrack/requirements.txt
napari==0.4.17
numpy==1.24.3
opencv_python==4.7.0.72
Pillow==9.5.0
scikit_image==0.19.3
scipy==1.9.3
tifffile==2023.4.12
tqdm==4.65.0